% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ma.R
\name{banded}
\alias{banded}
\title{Generates a banded covariance matrix and matrix squareroot
sig: value of kth band (starting with diagonal)
size of band is length(sig)}
\usage{
banded(p, sig)
}
\arguments{
\item{p}{dimension of covariance matrix}

\item{sig}{vector of values of Toeplitz matrix}
}
\description{
Generates a banded covariance matrix and matrix squareroot
sig: value of kth band (starting with diagonal)
size of band is length(sig)
}

