% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierband.R
\name{gpband}
\alias{gpband}
\title{Groupwise soft-thresholds subdiagonals by lam * w}
\usage{
gpband(R, lam, w = NULL)
}
\arguments{
\item{R}{p-by-p symmetric matrix}

\item{lam}{Non-negative penalty parameter.  Controls sparsity level.}

\item{w}{\code{(p-1)} vector of weights. Default, w[l]=sqrt(2 * l)}
}
\description{
Groupwise soft-thresholds subdiagonals by lam * w
}

