% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierband-package.R
\docType{package}
\name{hierband-package}
\alias{hierband-package}
\title{Convex banding of the covariance matrix using}
\description{
\code{hierband} is the R package implementing the convex banding approach to covariance estimation of Bien, Bunea, & Xiao (see full reference below).
}
\details{
The package is designed for situations in which there is a large number of variables that have a known ordering and in which it is believed that variables far apart in this ordering have little to no dependence.

It is called \code{hierband} (pronounced "hair band") because it makes use of a hierarchical group lasso penalty and provides a banded estimate of the covariance matrix.

Its main functions are \code{\link{hierband}}, \code{\link{hierband.path}}, \code{\link{hierband.cv}}.
}
\author{
Jacob Bien \email{jbien@cornell.edu}, Florentina Bunea, Luo Xiao
}
\references{
Bien, J., Bunea, F. Xiao, L. (2014) "Convex banding of the covariance matrix." Accepted for publication in JASA.
}

