% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierband.path.R
\name{lam.max.hierband}
\alias{lam.max.hierband}
\title{Computes the smallest lambda such that P=0.}
\usage{
lam.max.hierband(Sighat, ww)
}
\arguments{
\item{Sighat}{empirical covariance matrix}

\item{ww}{the diagonal of w}
}
\description{
Computes lambda_max, which is the smallest value of \code{lam} for which
\code{\link{hierband}} (with \code{delta=NULL}) gives a diagonal covariance matrix.
}

