% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{Covariance of an equal-weighted moving-average process}
\usage{
ma(p, K)
}
\arguments{
\item{p}{dimension of covariance matrix}

\item{K}{moving-average bandwidth}
}
\value{
Returns the covariance matrix, \code{Sig}, and the symmetric
square root, \code{A}, of this matrix.
}
\description{
Here, \code{Sig[j,k] = 0} if \code{|j-k|>K}
and \code{Sig[j,k] = 1 - |j-k| / K} otherwise.
}

