% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierband.cv.R
\name{MakeFolds}
\alias{MakeFolds}
\title{Make folds for cross validation}
\usage{
MakeFolds(n, nfolds)
}
\arguments{
\item{n}{sample size}

\item{nfolds}{number of folds}
}
\description{
Divides the indices \code{1:n} into \code{nfolds} random folds of about the same size.
}

