% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierband.R
\name{subdiag.thresh}
\alias{subdiag.thresh}
\title{Performs a single pass of BCD on a matrix R.}
\usage{
subdiag.thresh(R, lam, w = NULL)
}
\arguments{
\item{R}{p-by-p symmetric matrix}

\item{lam}{Non-negative penalty parameter.  Controls sparsity level.}

\item{w}{\code{(p-1)}-by-\code{(p-1)} lower-triangular matrix (above diagonal ignored).
\code{w[l,]} gives the \code{l} weights for g_l.
Defaults to \code{w[l,m]=sqrt(2 * l)/(l - m + 1)} for \code{m <= l}}
}
\description{
To solve the unconstrained problem, R is Sigmahat.
To solve constrained problem, R is the current partial residual (excluding A terms).
}

