% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierband.R
\name{subdiagonal.l2norms}
\alias{subdiagonal.l2norms}
\title{Compute the L2 norm of each subdiagonal of a symmetric matrix R.}
\usage{
subdiagonal.l2norms(R)
}
\arguments{
\item{R}{a symmetric matrix}
}
\description{
Compute the L2 norm of each subdiagonal of a symmetric matrix R.
}

